Be explicit about which getRaw() is called
Prevents infinite recursion due to implicit type casts when e. g.
loading the default volume in DeskVOX

We should really consider letting the frames counter be of type signed
size_t

+++ virvo/virvo/vvvoldesc.cpp
@@ -648,9 +648,9 @@ vvVolDesc::ErrorType vvVolDesc::mergeFrames()
 uint8_t* vvVolDesc::getRaw(int frame) const
 {
   if (frame == -1)
-    return getRaw(currentFrame);
+    return getRaw(static_cast< size_t >(currentFrame));
   else
-    return getRaw(frame);
+    return getRaw(static_cast< size_t >(frame));
 }
 
 //----------------------------------------------------------------------------
